android: PicturePublisher: reduced video lag by adding publishBacklogCount property and using it to reduce the amount of work in doInBackground method
Before this change, the video was always smooth but video in the web app could lag over a second behind the recorded frames because the tablet was falling behind in the upload process.  This commit fixes the problem by making it so only 1 task has any significant work in it at any given time so the tablet's frame uploads can be done as frequently as possible but frames will be dropped as needed to keep that delay under control.